+++ CHANGELOG.md
@@ -8,8 +8,9 @@
 
 **Fixes**:
 
-- Allow null custom log attributes without debug warning ([#1552](https://github.com/getsentry/sentry-native/pull/1552))
-
+- Allow null custom log attributes without debug warning. ([#1552](https://github.com/getsentry/sentry-native/pull/1552))
+- Introduce level-triggered wait flag for use in the batcher thread for logs and metrics. ([#1558](https://github.com/getsentry/sentry-native/pull/1558))
+ 
 ## 0.13.1
 
 **Features**: